chore: get pet profile function

+++ src/main/java/com/family/petmemory/entity/dto/WeightForm.java
@@ -4,6 +4,7 @@ import com.querydsl.core.annotations.QueryProjection;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,10 +14,10 @@ public class WeightForm {
 
     private LocalDate date;
 
-    private Float weight;
+    private BigDecimal weight;
 
     @QueryProjection
-    public WeightForm(Long id, LocalDate date, Float weight) {
+    public WeightForm(Long id, LocalDate date, BigDecimal weight) {
         this.id = id;
         this.date = date;
         this.weight = weight;
